pyproject.toml
[tool.poetry]
name = "skyranker"
version = "0.1.0"
description = ""
authors = ["Ryan Schuetzler <ryan@schuetzler.net>"]

[tool.poetry.dependencies]
python = "^3.8"
python-twitter = "^3.5"
python-dotenv = "^0.10.3"
sqlalchemy = "^1.3.12"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
